+++ ideas-service/src/main/kotlin/com/tyiu/ideas/service/SprintService.kt
@@ -56,7 +56,7 @@ class SprintService (val template: R2dbcEntityTemplate)
                         row.get("e_first_name", String::class.java),
                         row.get("e_last_name", String::class.java)
                     ),
-                    row.get("t_work_hour", Int::class.javaObjectType),
+                    row.get("t_work_hour", Double::class.javaObjectType),
                     row.get("t_start_date", LocalDate::class.java),
                     row.get("t_finish_date", LocalDate::class.java),
                     listOf(),
@@ -229,7 +229,7 @@ class SprintService (val template: R2dbcEntityTemplate)
                                 row.get("e_first_name", String::class.java),
                                 row.get("e_last_name", String::class.java)
                             ),
-                            row.get("t_work_hour", Int::class.javaObjectType),
+                            row.get("t_work_hour", Double::class.javaObjectType),
                             row.get("t_start_date", LocalDate::class.java),
                             row.get("t_finish_date", LocalDate::class.java),
                             listOf(),
