+++ ideas-service/src/main/kotlin/com/tyiu/ideas/service/TaskMovementLogService.kt
@@ -74,7 +74,7 @@ class TaskMovementLogService(val template: R2dbcEntityTemplate)
                                 row.get("te_first_name", String::class.java),
                                 row.get("te_last_name", String::class.java)
                             ),
-                            row.get("t_work_hour", Int::class.javaObjectType),
+                            row.get("t_work_hour", Double::class.javaObjectType),
                             row.get("t_start_date", LocalDate::class.java),
                             row.get("t_finish_date", LocalDate::class.java),
                             listOf(),
